code refactoring

+++ AndroidHomeApplication/app/src/main/java/com/example/androidhomeapplication/data/JsonGenre.kt
@@ -3,4 +3,4 @@ package com.example.androidhomeapplication.data
 import kotlinx.serialization.Serializable
 
 @Serializable
-internal class JsonGenre(val id: Int, val name: String)
+internal class JsonGenre(val id: Long, val name: String)
